fixed various warnings
git-svn-id: https://svn.salilab.org/imp/trunk@13030 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/isd/include/MolecularDynamicsMover.h
@@ -42,7 +42,7 @@ private:
 
 private:
   IMP::internal::OwnerPointer<MolecularDynamics> md_;
-  double nsteps_;
+  unsigned nsteps_;
   std::vector<std::vector<double> > coordinates_;
   std::vector<std::vector<double> > velocities_;
 };
